upd: little fix in daff_write, DFT content was writing sample rate as int32 instead of float32.
git-svn-id: https://svn.code.sf.net/p/opendaff/code/trunk@285 22735341-649f-4b50-8b46-d246c0249a88

+++ matlab/daff_write.m
@@ -929,7 +929,7 @@ function [] = daff_write( varargin )
         % Number of frequencies
         fwrite(fid, props.numDFTCoeffs, 'int32');
         fwrite(fid, props.transformSize, 'int32');
-        fwrite(fid, props.sampleRate, 'int32');
+        fwrite(fid, props.sampleRate, 'float32');
         fwrite(fid, props.globalPeak, 'float32');
     end
